fix: check media status has changed before modifying request
fix: refactor: code cleanup
